# sql: address correctness issues with the dependency digest